Refactor - Projectile Algo Dependencies
Removed need to specify initial frame upfront as android only provided the first frame when the algorithm was connected and running. Now a trade off has been made, where previous is always checked and initialised if empty, causing the very first compare to result in zero difference. This is fine.